refactor(site): extract AgentCreateForm from AgentsPage
