+++ site/src/pages/AgentsPage/AgentsPage.test.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from "@testing-library/react";
 import { beforeEach, describe, expect, it } from "vitest";
-import { emptyInputStorageKey, useEmptyStateDraft } from "./AgentsPage";
+import { emptyInputStorageKey, useEmptyStateDraft } from "./AgentCreateForm";
 
 describe("useEmptyStateDraft", () => {
 	beforeEach(() => {
